Finalise headers methods
